refactor: restructure Bookmark module
Bookmark now has two subpackages: api & impl. Code is currently in impl.
