+++ feature/bookmark/api/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    alias(libs.plugins.kite.android.feature.api)
+}
+
+android {
+    namespace = "com.potatosheep.kite.feature.bookmark.api"
+}
+
+dependencies {
+    api(projects.core.navigation)
+}
