+++ feature/bookmark/impl/src/main/AndroidManifest.xml
