+++ settings.gradle.kts
@@ -48,12 +48,13 @@ include(":feature:settings")
 include(":core:datastore")
 include(":core:datastore-proto")
 include(":core:database")
-include(":feature:bookmark")
+include(":feature:bookmark:impl")
+include(":feature:bookmark:api")
 include(":feature:exception")
 include(":feature:onboarding")
 include(":feature:about:impl")
+include(":feature:about:api")
 include(":core:media")
 include(":core:notification")
 include(":core:translation")
 include(":core:navigation")
-include(":feature:about:api")
